+++ subvertpy/wc.c
@@ -899,7 +899,7 @@ static PyObject *committed_queue_queue(CommittedQueueObject *self, PyObject *arg
 	char *md5_digest = NULL, *sha1_digest = NULL;
 	bool recurse = false;
 	apr_array_header_t *wcprop_changes;
-	int md5_digest_len, sha1_digest_len;
+	Py_ssize_t md5_digest_len, sha1_digest_len;
 #if ONLY_SINCE_SVN(1, 7)
 	svn_wc_context_t *context = NULL;
 #endif
